+++ src/scripts/ci.clj
@@ -42,6 +42,12 @@
   (shell "./gradlew build publishToMavenLocal")
   (fs/delete-tree "src/main/resources/com/github/ericdallo/clj4intellij"))
 
+(defn build [& _]
+  (shell "./gradlew clean")
+  (fs/copy-tree "src/main/clojure/" "src/main/resources/")
+  (shell "./gradlew build")
+  (fs/delete-tree "src/main/resources/com/github/ericdallo/clj4intellij"))
+
 (defn deploy [& _]
   (shell "./gradlew clean")
   (fs/copy-tree "src/main/clojure/" "src/main/resources/")